Follow the measured Week 2 decode bottleneck
Move fused model kernels before decode attention and make the cumulative checkpoints, tests, benchmark labels, and Week 3 prerequisites follow the same order.
