+++ benches/bench_course_progression.py
@@ -54,32 +54,32 @@ WEEK2_VARIANTS = (
         ("--week2-checkpoint", "quantized-matvec"),
     ),
     Variant(
-        "week2-decode-attention",
-        "2.4 Decode attention",
-        "ref",
-        "week2",
-        ("--week2-checkpoint", "decode-attention"),
-    ),
-    Variant(
         "week2-rmsnorm",
-        "2.5 Fast RMSNorm",
+        "2.4 Fast RMSNorm",
         "ref",
         "week2",
         ("--week2-checkpoint", "rmsnorm"),
     ),
     Variant(
         "week2-rope",
-        "2.5 + Fast RoPE",
+        "2.4 + Fast RoPE",
         "ref",
         "week2",
         ("--week2-checkpoint", "rope"),
     ),
     Variant(
         "week2-swiglu",
-        "2.5 + Fused SwiGLU",
+        "2.4 + Fused SwiGLU",
         "ref",
         "week2",
         ("--week2-checkpoint", "swiglu"),
+    ),
+    Variant(
+        "week2-decode-attention",
+        "2.5 Decode attention",
+        "ref",
+        "week2",
+        ("--week2-checkpoint", "decode-attention"),
     ),
     Variant(
         "week2-simd-matmul",
