+++ book/src/SUMMARY.md
@@ -17,8 +17,8 @@
     - [🚧 KV Cache](./week2-01-kv-cache.md)
     - [🚧 Benchmark and Profile](./week2-02-benchmarking.md)
     - [🚧 Quantized Matvec](./week2-03-quantized-matvec.md)
-    - [🚧 Fused Decode Attention](./week2-04-decode-attention.md)
-    - [🚧 Fused Model Kernels](./week2-05-fast-kernels.md)
+    - [🚧 Fused Model Kernels](./week2-04-fused-model-kernels.md)
+    - [🚧 Fused Decode Attention](./week2-05-decode-attention.md)
     - [🚧 SIMD-Matrix Prefill](./week2-06-simd-matrix-prefill.md)
     - [🚧 Split-K Prefill](./week2-07-split-k-prefill.md)
 - [🚧 Week 3: Build a Mini vLLM](./week3-overview.md)
