+++ book/src/glossary.md
@@ -15,9 +15,9 @@
 - [Week 1 Qwen3 Model](./week1-05-qwen3-model.md)
 - [dequantize_linear](./week1-05-qwen3-model.md)
 - [Quantized Matmul](./week2-03-quantized-matvec.md)
-- [Fused Model Kernels](./week2-05-fast-kernels.md)
+- [Fused Model Kernels](./week2-04-fused-model-kernels.md)
 - [KV Cache](./week2-01-kv-cache.md)
-- [Fused Decode Attention](./week2-04-decode-attention.md)
+- [Fused Decode Attention](./week2-05-decode-attention.md)
 - [SIMD-Matrix Prefill](./week2-06-simd-matrix-prefill.md)
 - [Flash Attention](./week3-05-flash-attention.md)
 - [Paged Attention](./week3-04-paged-attention-part2.md)
