+++ book/src/week2-03-quantized-matvec.md
@@ -616,12 +616,14 @@ pdm run bench --solution tiny_llm --loader week2 \
 Run the same command with `--solution tiny_llm_ref` to compare it with the
 reference solution.
 
-Compare this result with the Day 1 `kv-cache` row. Do not start the decode
-attention chapter until the complete model uses packed weights and the
-end-to-end number has been recorded. The vanilla matrix product remains
-callable as a correctness oracle, but only the SIMD matvec is integrated into
-decode. Reprofile the quantized checkpoint across increasing cached context.
-Day 4 is justified when attention's share grows with context after projection
-traffic has fallen; otherwise keep tuning the measured matvec bottleneck.
+Compare this result with the Day 1 `kv-cache` row. Do not start the fused model
+kernels until the complete model uses packed weights and the end-to-end number
+has been recorded. The vanilla matrix product remains callable as a correctness
+oracle, but only the SIMD matvec is integrated into decode. Reprofile the
+quantized checkpoint and compare each optimized projection with MLX at the same
+Qwen3-4B shape. When projection latency is close to that external denominator
+but normalization, position, and activation still account for the largest
+removable gap, continue to Day 4. Otherwise keep tuning the measured matvec
+bottleneck.
 
 {{#include copyright.md}}
